Split out getPostsForQuery
